requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile
#
amqp==5.2.0
    # via kombu
asgiref==3.8.1
    # via django
asttokens==2.4.1
    # via stack-data
async-timeout==4.0.3
    # via redis
attrs==24.1.0
    # via
    #   jsonschema
    #   referencing
billiard==3.6.4.0
    # via celery
black==24.8.0
    # via -r requirements.in
build==1.2.1
    # via pip-tools
celery==5.0.2
    # via
    #   -r requirements.in
    #   django-celery-beat
cffi==1.17.0
    # via cryptography
click==8.1.7
    # via
    #   black
    #   celery
    #   click-didyoumean
    #   click-repl
    #   pip-tools
click-didyoumean==0.3.1
    # via celery
click-repl==0.3.0
    # via celery
colorama==0.4.6
    # via
    #   build
    #   click
    #   ipython
cryptography==43.0.0
    # via -r requirements.in
decorator==5.1.1
    # via ipython
django==3.2.25
    # via
    #   -r requirements.in
    #   django-celery-beat
    #   django-filter
    #   django-redis
    #   django-timezone-field
    #   djangorestframework
    #   djangorestframework-simplejwt
    #   drf-yasg
django-celery-beat==2.2.1
    # via -r requirements.in
django-filter==23.5
    # via -r requirements.in
django-mail-templated==2.6.5
    # via -r requirements.in
django-redis==5.4.0
    # via -r requirements.in
django-timezone-field==4.2.3
    # via django-celery-beat
djangorestframework==3.15.1
    # via
    #   -r requirements.in
    #   djangorestframework-simplejwt
    #   drf-yasg
djangorestframework-simplejwt==5.3.1
    # via -r requirements.in
drf-yasg[validation]==1.21.7
    # via -r requirements.in
envparse==0.2.0
    # via -r requirements.in
exceptiongroup==1.2.2
    # via ipython
executing==2.0.1
    # via stack-data
faker==26.1.0
    # via -r requirements.in
flake8==7.1.1
    # via -r requirements.in
gunicorn==22.0.0
    # via -r requirements.in
importlib-resources==6.4.0
    # via swagger-spec-validator
inflection==0.5.1
    # via drf-yasg
ipython==8.9.0
    # via -r requirements.in
jedi==0.19.1
    # via ipython
jsonschema==4.23.0
    # via swagger-spec-validator
jsonschema-specifications==2023.12.1
    # via jsonschema
kombu==5.3.7
    # via celery
markdown==3.6
    # via -r requirements.in
matplotlib-inline==0.1.7
    # via ipython
mccabe==0.7.0
    # via flake8
mypy-extensions==1.0.0
    # via black
packaging==24.1
    # via
    #   black
    #   build
    #   drf-yasg
    #   gunicorn
parso==0.8.4
    # via jedi
pathspec==0.12.1
    # via black
pillow==10.4.0
    # via -r requirements.in
pip-tools==7.4.1
    # via -r requirements.in
platformdirs==4.2.2
    # via black
prompt-toolkit==3.0.47
    # via
    #   click-repl
    #   ipython
psycopg2-binary==2.9.9
    # via -r requirements.in
pure-eval==0.2.3
    # via stack-data
pycodestyle==2.12.1
    # via flake8
pycparser==2.22
    # via cffi
pyflakes==3.2.0
    # via flake8
pygments==2.18.0
    # via ipython
pyjwt==2.9.0
    # via djangorestframework-simplejwt
pyproject-hooks==1.1.0
    # via
    #   build
    #   pip-tools
python-crontab==3.2.0
    # via django-celery-beat
python-dateutil==2.9.0.post0
    # via
    #   faker
    #   python-crontab
python-decouple==3.8
    # via -r requirements.in
pytz==2024.1
    # via
    #   celery
    #   django
    #   django-timezone-field
    #   drf-yasg
pyyaml==6.0.1
    # via
    #   drf-yasg
    #   swagger-spec-validator
redis==5.0.8
    # via
    #   -r requirements.in
    #   django-redis
referencing==0.35.1
    # via
    #   jsonschema
    #   jsonschema-specifications
rpds-py==0.19.1
    # via
    #   jsonschema
    #   referencing
six==1.16.0
    # via
    #   asttokens
    #   python-dateutil
sqlparse==0.5.1
    # via django
stack-data==0.6.3
    # via ipython
swagger-spec-validator==3.0.4
    # via drf-yasg
tomli==2.0.1
    # via
    #   black
    #   build
    #   pip-tools
traitlets==5.14.3
    # via
    #   ipython
    #   matplotlib-inline
typing-extensions==4.12.2
    # via
    #   asgiref
    #   black
    #   ipython
    #   swagger-spec-validator
uritemplate==4.1.1
    # via drf-yasg
vine==5.1.0
    # via
    #   amqp
    #   celery
    #   kombu
wcwidth==0.2.13
    # via prompt-toolkit
wheel==0.44.0
    # via pip-tools

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
